# docs/handoff/ups-expresosapp-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/request-code": {
      "post": {
        "operationId": "AuthController_requestCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Verification code sent"
          },
          "429": {
            "description": "Too many requests. Try again later."
          }
        },
        "summary": "Request an OTP verification code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-code": {
      "post": {
        "operationId": "AuthController_verifyCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tokens generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensDto"
                }
              }
            }
          }
        },
        "summary": "Verify OTP and obtain access/refresh tokens",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tokens refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokensDto"
                }
              }
            }
          }
        },
        "summary": "Refresh access token using a refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        },
        "summary": "Logout and revoke session",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current authenticated user",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/admin/drivers": {
      "post": {
        "operationId": "DriversController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Driver created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new driver",
        "tags": [
          "Admin Drivers"
        ]
      },
      "get": {
        "operationId": "DriversController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "description": "Page number (1-based)",
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "description": "Items per page",
            "required": false,
            "name": "limit",
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of drivers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverPaginatedResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all drivers with pagination",
        "tags": [
          "Admin Drivers"
        ]
      }
    },
    "/admin/drivers/{id}": {
      "get": {
        "operationId": "DriversController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Driver ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Driver details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Driver not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get driver details",
        "tags": [
          "Admin Drivers"
        ]
      },
      "patch": {
        "operationId": "DriversController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Driver ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Driver updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Driver not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a driver",
        "tags": [
          "Admin Drivers"
        ]
      }
    },
...
  },
  "info": {
    "title": "UPS ExpresosApp API",
    "description": "Backend API for UPS ExpresosApp MVP - Institutional transport management system for Universidad Politecnica Salesiana.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RequestCodeDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Institutional email address used to request OTP verification",
            "example": "student@est.ups.edu.ec"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyCodeDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Institutional email address",
            "example": "student@est.ups.edu.ec"
          },
          "code": {
            "type": "string",
            "description": "OTP verification code received via email",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "email": {
            "type": "string",
            "example": "student@est.ups.edu.ec"
          },
          "name": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "STUDENT",
              "ADMIN",
              "SUPER_ADMIN",
              "DRIVER"
            ],
            "example": "STUDENT"
          },
          "emailVerified": {
            "type": "boolean",
            "example": true
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "emailVerified",
          "isActive"
        ]
      },
      "AuthTokensDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token"
          },
          "refreshToken": {
            "type": "string",
            "description": "Refresh token for obtaining new access tokens"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token obtained from login or refresh response",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token to revoke. If not provided, attempts to revoke by session context.",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
...
      "CreateDriverDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Driver name",
            "example": "Juan Pérez"
          },
          "phone": {
            "type": "string",
            "description": "Driver phone number",
            "example": "+593991234567"
          },
          "licenseNumber": {
            "type": "string",
            "description": "Driver license number",
            "example": "L123456789"
          },
          "status": {
            "type": "string",
            "description": "Driver status",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "default": "ACTIVE"
          },
          "assignedVehicleId": {
            "type": "string",
            "description": "Assigned vehicle ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "assignedRouteId": {
            "type": "string",
            "description": "Assigned route ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        },
        "required": [
          "name"
        ]
      },
      "DriverResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "example": "Juan Pérez"
          },
          "phone": {
            "type": "object",
            "example": "+593991234567",
            "nullable": true
          },
          "licenseNumber": {
            "type": "object",
            "example": "L123456789",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "example": "ACTIVE"
          },
          "assignedVehicleId": {
            "type": "object",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "nullable": true
          },
          "assignedRouteId": {
            "type": "object",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "licenseNumber",
          "status",
          "assignedVehicleId",
          "assignedRouteId",
          "createdAt",
          "updatedAt"
        ]
      },
      "DriverPaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DriverResponseDto"
            }
          },
          "meta": {
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateDriverDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Driver name",
            "example": "Juan Pérez"
          },
          "phone": {
            "type": "string",
            "description": "Driver phone number",
            "example": "+593991234567"
          },
          "licenseNumber": {
            "type": "string",
            "description": "Driver license number",
            "example": "L123456789"
          },
          "status": {
            "type": "string",
            "description": "Driver status",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "default": "ACTIVE"
          },
          "assignedVehicleId": {
            "type": "string",
            "description": "Assigned vehicle ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "assignedRouteId": {
            "type": "string",
            "description": "Assigned route ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        }
      },
...
    }
  }
}